requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.48
alembic==1.14.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
python-dotenv==1.0.1
slowapi==0.1.9
pandas==2.2.2
openpyxl==3.1.5
bleach==6.3.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
pytest==8.3.5
pydantic[email]==2.10.4
